requirements.txt
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.9.0
promise==2.3
PyJWT==1.7.1
pytz==2019.3
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
